First draft of report with introduction, and Data Exploratory Analysis
I've included the first component of the Spam Detector - this just loads the corpus into a usable list of objects, and the Jupiter Notebook for the actual analysis. The Jupiter Notebook is probably a better place to start with coding as the code here is more up to date. Will find a way to actually upload the corpus momentarily as I don't know how to upload an entire directory.